requirements.txt
snowflake-snowpark-python
behave
pandas
